📝 style: Funnel, Step 컴포넌트를 한 파일에 colocation
