+++ src/components/Funnel/createFunnel.ts
@@ -1,6 +1,5 @@
 import useFunnel from '@/hooks/useFunnel';
-import Funnel from './Funnel';
-import Step from './Step';
+import Funnel, { Step } from '.';
 
 /**
  * 퍼널 컴포넌트와 훅을 이용하는데 있어서 스텝의 순서를 자동으로 제네릭에 의해 관리할 수 있도록 도와주는 함수
